Extract m_template definition to template.h
