+++ template.h
@@ -1,0 +1,29 @@
+#ifndef V8GO_TEMPLATE_H
+#define V8GO_TEMPLATE_H
+#ifdef __cplusplus
+
+namespace v8 {
+class Isolate;
+class Template;
+}  // namespace v8
+
+#include "deps/include/v8-persistent-handle.h"
+
+struct m_template {
+  v8::Isolate* iso;
+  v8::Persistent<v8::Template> ptr;
+};
+
+extern "C" {
+#else
+
+typedef struct m_template m_template;
+
+#endif
+
+typedef m_template* TemplatePtr;
+
+#ifdef __cplusplus
+}  // extern "C"
+#endif
+#endif
